refactor(rating-info): replace average totalPoint with integer RatingTotal sum

+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/adapter/in/web/request/PostRatingRequest.java
@@ -1,18 +1,12 @@
 package vn.uit.edu.msshop.rating.adapter.in.web.request;
 
-
 import java.util.UUID;
 
 public record PostRatingRequest(
-   
-    
-    String content,
-    UUID productId,
-    UUID userId,
-    String userName,
-    String userAvatar,
-    float ratingPoint
-
-) {
-
+        String content,
+        UUID productId,
+        UUID userId,
+        String userName,
+        String userAvatar,
+        byte ratingPoint) {
 }
