+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/adapter/in/web/request/UpdateRatingRequest.java
@@ -5,11 +5,7 @@ import java.util.UUID;
 import vn.uit.edu.msshop.rating.adapter.in.web.request.common.ChangeRequest;
 
 public record UpdateRatingRequest(
-    UUID id,
-    ChangeRequest<String> content,
-    ChangeRequest<Float> ratingPoint
-    
-
-) {
-    
+        UUID id,
+        ChangeRequest<String> content,
+        ChangeRequest<Byte> ratingPoint) {
 }
