+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/application/dto/query/RatingView.java
@@ -2,6 +2,15 @@ package vn.uit.edu.msshop.rating.application.dto.query;
 
 import java.util.UUID;
 
-public record RatingView(UUID ratingId, String content, String mediaURL, String mediaType, String mediaPublicId, UUID productId, UUID userId, String userName, String userAvatar, float ratingPoint) {
-
+public record RatingView(
+        UUID ratingId,
+        String content,
+        String mediaURL,
+        String mediaType,
+        String mediaPublicId,
+        UUID productId,
+        UUID userId,
+        String userName,
+        String userAvatar,
+        byte ratingPoint) {
 }
